small css mistake, thousand's separator
I've done the first pull of the code and I've decided not to implement a thousand's separator.
And all this week, every day, this "space char" were in my head. I've done some researches and, obviously, it is more comfortable with it.
Thus I've implemented it. Small function called "addSpaces" and .replace for calculation part.
+ option's style review